# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShieldCap API — confidential real estate tokenization platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PropertyStatus = typeof PropertyStatus[keyof typeof PropertyStatus];


export const PropertyStatus = {
  active: 'active',
  paused: 'paused',
  sold: 'sold',
} as const;

export interface Property {
  id: number;
  name: string;
  location: string;
  description: string;
  valueUsd: number;
  totalShares: number;
  pricePerShare: number;
  imageUrl: string;
  status: PropertyStatus;
}

export interface PropertyMetrics {
  totalSharesIssued: number;
  totalRevenueDistributedUsd: number;
  totalTransactions: number;
  totalDividendRounds: number;
  investorCount: number;
  propertyValueUsd: number;
  monthlyRevenueUsd: number;
}

export type ContractConfigAbiItem = { [key: string]: unknown };

export interface ContractConfig {
  contractAddress: string;
  paymentTokenAddress?: string | null;
  paymentTokenSymbol?: string | null;
  paymentTokenDecimals?: number | null;
  networkId: number;
  networkName: string;
  abi: ContractConfigAbiItem[];
}

export type TransactionEventEventType = typeof TransactionEventEventType[keyof typeof TransactionEventEventType];


export const TransactionEventEventType = {
  Purchase: 'Purchase',
  Transfer: 'Transfer',
  DividendDistribution: 'DividendDistribution',
  CapRejected: 'CapRejected',
} as const;

export interface TransactionEvent {
  id: number;
  txHash: string;
  eventType: TransactionEventEventType;
  timestamp: string;
  blockNumber: number;
  /** @nullable */
  walletAddress?: string | null;
}

export type DividendRoundStatus = typeof DividendRoundStatus[keyof typeof DividendRoundStatus];


export const DividendRoundStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  failed: 'failed',
} as const;

export interface DividendRound {
  id: number;
  roundNumber: number;
  totalDistributedUsd: number;
  revenueUsd: number;
  distributedAt: string;
  /** @nullable */
  txHash: string | null;
  status: DividendRoundStatus;
}

export interface DividendInput {
  revenueUsd: number;
  txHash: string;
}

export interface InvestorProfile {
  id: number;
  walletAddress: string;
  registeredAt: string;
  transactionCount: number;
  dividendRoundsParticipated: number;
  /** @nullable */
  displayName?: string | null;
}

export interface InvestorInput {
  walletAddress: string;
  displayName?: string;
}

export type InvestorDividendRecordStatus = typeof InvestorDividendRecordStatus[keyof typeof InvestorDividendRecordStatus];


export const InvestorDividendRecordStatus = {
  eligible: 'eligible',
  excluded: 'excluded',
} as const;

export interface InvestorDividendRecord {
  id: number;
  roundNumber: number;
  distributedAt: string;
  status: InvestorDividendRecordStatus;
}